+++ CHANGES.rst
@@ -6,6 +6,10 @@ Change Log
 
 Released on XXX, 2013
 
+* Removed ``simpletree`` from the package. The default tree builder is
+  now ``etree`` (using the ``xml.etree.ElementTree/cElementTree``
+  implementation).
+
 
 0.95
 ~~~~